+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+# 0.1.1 (2018-06-17)
+
+* Bugfixed options initialization.
+* First try at downsizing the mutant's canvas down to the bounds
+  corresponding to the region reported by the mutant.
+
+
 # 0.1.0 (2018-06-17)
 
-* Initial release
+* Initial release.